fix not changing transmit power level to 0 and always setting channel to 0

+++ si4432.h
@@ -282,7 +282,7 @@ protected:
 
 	float _freqCarrier;
 	float _kbps;
-	uint8_t _freqChannel;
+	uint16_t _freqChannel;
 
 	ModulationType _modulationType;
